Fix Auth Model bug, Update Models
the auth model was causing a bug in the application because
the name of the collection, auth, was conflicting with mongo's
db.auth() functionality. the fix was to explicitly state a
collection name that shouldn't introduce any conflicts

Session model was also introducing bugs 'cause it did not inherit
from the Document class.